+++ host/bindings/bindings.h
@@ -8,6 +8,7 @@
 #define MCUJS_BINDINGS_H
 
 #include <stdbool.h>
+#include "jerryscript.h"
 
 /*
  * Register console bindings
@@ -20,6 +21,16 @@ void js_bind_console(void);
  * GPIO.init(), GPIO.set(), GPIO.get(), GPIO.toggle()
  */
 void js_bind_gpio(void);
+
+/*
+ * Create GPIO module object
+ */
+jerry_value_t js_create_gpio_module(void);
+
+/*
+ * Create GPIO module object
+ */
+jerry_value_t js_create_gpio_module(void);
 
 /*
  * Register timer bindings
@@ -40,10 +51,20 @@ bool js_timers_process(void);
 void js_bind_pwm(void);
 
 /*
+ * Create PWM module object
+ */
+jerry_value_t js_create_pwm_module(void);
+
+/*
  * Register I2C bindings
  * I2C.init(), I2C.write(), I2C.read()
  */
 void js_bind_i2c(void);
+
+/*
+ * Create I2C module object
+ */
+jerry_value_t js_create_i2c_module(void);
 
 /*
  * Register SPI bindings
@@ -52,10 +73,20 @@ void js_bind_i2c(void);
 void js_bind_spi(void);
 
 /*
+ * Create SPI module object
+ */
+jerry_value_t js_create_spi_module(void);
+
+/*
  * Register ADC bindings
  * adc.readPin(), adc.readChannel(), adc.readTempC(), adc.TEMP
  */
 void js_bind_adc(void);
+
+/*
+ * Create ADC module object
+ */
+jerry_value_t js_create_adc_module(void);
 
 /*
  * Register board bindings
@@ -68,6 +99,11 @@ void js_bind_board(void);
  * fs.readFileSync(), fs.writeFileSync(), fs.existsSync(), etc.
  */
 void js_bind_fs(void);
+
+/*
+ * Create filesystem module object
+ */
+jerry_value_t js_create_fs_module(void);
 
 /*
  * Register process bindings (Node.js-compatible)